+++ tdrs-backend/tdpservice/parsers/test/test_validators.py
@@ -3,6 +3,7 @@
 import pytest
 import logging
 from .. import validators
+from ..case_consistency_validator import CaseConsistencyValidator
 from .. import schema_defs, util
 from tdpservice.parsers.test.factories import TanfT1Factory, TanfT2Factory, TanfT3Factory, TanfT5Factory, TanfT6Factory
 from tdpservice.parsers.test.factories import SSPM5Factory
@@ -1234,8 +1235,8 @@ class TestM5Cat3Validators(TestCat3ValidatorsBase):
                           ['FAMILY_AFFILIATION', 'REC_FEDERAL_DISABILITY'])
 
 
-class TestCat4Validator:
-    """Test category four validators."""
+class TestCaseConsistencyValidator:
+    """Test case consistency (cat4) validators."""
 
     def parse_header(self, datafile):
         """Parse datafile header into header object."""
@@ -1259,6 +1260,14 @@ class TestCat4Validator:
         return [t1, t2, t3, t3_1]
 
     @pytest.fixture
+    def tanf_s1_schemas(self):
+        """Return group of TANF Section 1 schemas."""
+        s1 = schema_defs.tanf.t1.schemas[0]
+        s2 = schema_defs.tanf.t2.schemas[0]
+        s3 = schema_defs.tanf.t3.schemas[0]
+        return [s1, s2, s3, s3]
+
+    @pytest.fixture
     def small_correct_file(self, stt_user, stt):
         """Fixture for small_correct_file."""
         return util.create_test_datafile('small_correct_file.txt', stt_user, stt)
@@ -1274,70 +1283,70 @@ class TestCat4Validator:
         return header
 
     @pytest.mark.django_db
-    def test_add_record(self, small_correct_file_header, small_correct_file, tanf_s1_records):
+    def test_add_record(self, small_correct_file_header, small_correct_file, tanf_s1_records, tanf_s1_schemas):
         """Test add_record logic."""
-        cat_four_validator = validators.CatFourValidator(small_correct_file_header,
-                                                         util.make_generate_parser_error(small_correct_file, None))
+        case_consistency_validator = CaseConsistencyValidator(small_correct_file_header,
+                                                              util.make_generate_parser_error(small_correct_file, None))
 
-        for record in tanf_s1_records:
-            cat_four_validator.add_record(record, True)
+        for record, schema in zip(tanf_s1_records, tanf_s1_schemas):
+            case_consistency_validator.add_record(record, schema, True)
 
-        assert cat_four_validator.has_validated is False
-        assert cat_four_validator.case_has_errors is True
-        assert len(cat_four_validator.records) == 4
-        assert cat_four_validator.total_cases_cached == 0
-        assert cat_four_validator.total_cases_validated == 0
+        assert case_consistency_validator.has_validated is False
+        assert case_consistency_validator.case_has_errors is True
+        assert len(case_consistency_validator.records) == 4
+        assert case_consistency_validator.total_cases_cached == 0
+        assert case_consistency_validator.total_cases_validated == 0
 
         # Add record with different case number to proc validation again and start caching a new case.
         t1 = TanfT1Factory.create()
         t1.CASE_NUMBER = 2
-        cat_four_validator.add_record(t1, False)
-        assert cat_four_validator.has_validated is False
-        assert cat_four_validator.case_has_errors is False
-        assert len(cat_four_validator.records) == 1
-        assert cat_four_validator.total_cases_cached == 1
-        assert cat_four_validator.total_cases_validated == 0
+        case_consistency_validator.add_record(t1, tanf_s1_schemas[0], False)
+        assert case_consistency_validator.has_validated is False
+        assert case_consistency_validator.case_has_errors is False
+        assert len(case_consistency_validator.records) == 1
+        assert case_consistency_validator.total_cases_cached == 1
+        assert case_consistency_validator.total_cases_validated == 0
 
         # Complete the case to proc validation and verify that it occured. Even if the next case has errors.
         t2 = TanfT2Factory.create()
         t3 = TanfT3Factory.create()
         t2.CASE_NUMBER = 2
         t3.CASE_NUMBER = 2
-        cat_four_validator.add_record(t2, False)
-        cat_four_validator.add_record(t3, False)
-        assert cat_four_validator.case_has_errors is False
+        case_consistency_validator.add_record(t2, tanf_s1_schemas[1], False)
+        case_consistency_validator.add_record(t3, tanf_s1_schemas[2], False)
+        assert case_consistency_validator.case_has_errors is False
 
-        cat_four_validator.add_record(tanf_s1_records[0], True)
+        case_consistency_validator.add_record(tanf_s1_records[0], tanf_s1_schemas[0], True)
 
-        assert cat_four_validator.has_validated is True
-        assert cat_four_validator.case_has_errors is True
-        assert len(cat_four_validator.records) == 1
-        assert cat_four_validator.total_cases_cached == 2
-        assert cat_four_validator.total_cases_validated == 1
+        assert case_consistency_validator.has_validated is True
+        assert case_consistency_validator.case_has_errors is True
+        assert len(case_consistency_validator.records) == 1
+        assert case_consistency_validator.total_cases_cached == 2
+        assert case_consistency_validator.total_cases_validated == 1
 
     @pytest.mark.django_db
-    def test_section1_fail(self, small_correct_file_header, small_correct_file, tanf_s1_records):
+    def test_section1_fail(self, small_correct_file_header, small_correct_file, tanf_s1_records, tanf_s1_schemas):
         """Test TANF Section 1 records RPT_MONTH_YEAR don't align with header year and quarter."""
-        cat_four_validator = validators.CatFourValidator(small_correct_file_header,
-                                                         util.make_generate_parser_error(small_correct_file, None))
+        case_consistency_validator = CaseConsistencyValidator(small_correct_file_header,
+                                                              util.make_generate_parser_error(small_correct_file, None))
 
-        for record in tanf_s1_records:
-            cat_four_validator.add_record(record, False)
+        for record, schema in zip(tanf_s1_records, tanf_s1_schemas):
+            case_consistency_validator.add_record(record, schema, False)
 
-        num_errors = cat_four_validator.validate()
+        num_errors = case_consistency_validator.validate()
 
         assert 4 == num_errors
 
     @pytest.mark.django_db
-    def test_section1_pass(self, small_correct_file_header, small_correct_file, tanf_s1_records):
+    def test_section1_pass(self, small_correct_file_header, small_correct_file, tanf_s1_records, tanf_s1_schemas):
         """Test TANF Section 1 records RPT_MONTH_YEAR do align with header year and quarter."""
-        cat_four_validator = validators.CatFourValidator(small_correct_file_header,
-                                                         util.make_generate_parser_error(small_correct_file, None))
+        case_consistency_validator = CaseConsistencyValidator(small_correct_file_header,
+                                                              util.make_generate_parser_error(small_correct_file, None))
 
-        for record in tanf_s1_records:
+        for record, schema in zip(tanf_s1_records, tanf_s1_schemas):
             record.RPT_MONTH_YEAR = 202010
-            cat_four_validator.add_record(record, False)
+            case_consistency_validator.add_record(record, schema, False)
 
-        num_errors = cat_four_validator.validate()
+        num_errors = case_consistency_validator.validate()
 
         assert 0 == num_errors
